Enhance game configuration and save system: Update README to outline critical blockers and major system gaps for early access launch. Implement comprehensive save/load functionality, including auto-save settings and daily cycle management. Refactor GameEngine and DailySystemManager for improved game state handling and player experience. Introduce settings menu for save management and daily cycle configuration in MainStreetLocation. These changes aim to streamline gameplay and enhance user engagement.

+++ Scripts/Systems/QuestSystem.cs
@@ -77,7 +77,7 @@ public partial class QuestSystem : Node
         // Claim the quest
         foundQuest.Occupier = player.Name2;
         foundQuest.OccupierRace = player.Race;
-        foundQuest.OccupierSex = (byte)player.Sex;
+                        foundQuest.OccupierSex = (byte)((int)player.Sex);
         foundQuest.OccupiedDays = 0;
         
         // Track in player list
